+++ Problems/P/P_NFA/Visualizations/NFAVisualization.cs
@@ -1,12 +1,12 @@
 using API.Interfaces;
 using API.Interfaces.JSON_Objects;
-using API.Problems.NPComplete.NPC_NFA;
+using API.Problems.P.P_NFA;
 using API.Interfaces.JSON_Objects.Graphs;
 using System.Text.Json;
 using API.Interfaces.Graphs.GraphParser;
-using API.Problems.NPComplete.NPC_NFA.Solvers;
+using API.Problems.P.P_NFA.Solvers;
 
-namespace API.Problems.NPComplete.NPC_NFA.Visualizations;
+namespace API.Problems.P.P_NFA.Visualizations;
 
 class NFAVisualization : IVisualization<NFA>
 {
